+++ inc/services/class.bitly.service.php
@@ -20,80 +20,51 @@ class bitlyKutrlService extends kutrlService
     protected $config = [
         'id'         => 'bitly',
         'name'        => 'bit.ly',
-        'home'        => 'http://bit.ly',
+        'home'        => 'https://bit.ly',
 
-        'url_api'     => 'http://api.bit.ly/v3/',
-        'url_base'    => 'http://bit.ly/',
-        'url_min_len' => 25
+        'url_api'     => 'https://api-ssl.bitly.com/v4/',
+        'url_base'    => 'https://bit.ly/',
+        'url_min_len' => 25,
+
+        'allow_protocols'    => ['http://', 'https://'],
     ];
 
     private $args = [
-        'format'  => 'xml',
-        'login'   => '',
-        'apiKey'  => '',
-        'history' => 0
+        'apiKey'  => ''
     ];
 
     protected function init()
     {
-        $this->args['login']   = $this->settings->kutrl_srv_bitly_login;
         $this->args['apiKey']  = $this->settings->kutrl_srv_bitly_apikey;
-        $this->args['history'] = $this->settings->kutrl_srv_bitly_history ? 1 : 0;
     }
 
     public function saveSettings()
     {
-        $this->settings->put('kutrl_srv_bitly_login', $_POST['kutrl_srv_bitly_login']);
         $this->settings->put('kutrl_srv_bitly_apikey', $_POST['kutrl_srv_bitly_apikey']);
-        $this->settings->put('kutrl_srv_bitly_history', isset($_POST['kutrl_srv_bitly_history']));
     }
 
     public function settingsForm()
     {
         echo 
-        '<p><label class="classic">' . __('Login:') . '<br />' . 
-        form::field(['kutrl_srv_bitly_login'], 50, 255, $this->settings->kutrl_srv_bitly_login) . 
-        '</label></p>' . 
-        '<p class="form-note">' . 
-        sprintf(__('This is your login to sign up to %s'), $this->config['name']) . 
-        '</p>' . 
         '<p><label class="classic">' . __('API Key:') . '<br />' . 
         form::field(['kutrl_srv_bitly_apikey'], 50, 255, $this->settings->kutrl_srv_bitly_apikey) . 
         '</label></p>' . 
         '<p class="form-note">' . 
         sprintf(__('This is your personnal %s API key. You can find it on your account page.'), $this->config['name']) . 
-        '</p>' . 
-        '<p><label class="classic">' . 
-        form::checkbox(['kutrl_srv_bitly_history'], '1', $this->settings->kutrl_srv_bitly_history) . ' ' . 
-        __('Publish history') . 
-        '</label></p>' . 
-        '<p class="form-note">' . 
-        __('This publish all short links on your bit.ly public page.') . 
         '</p>';
     }
 
     public function testService()
     {
-        if (empty($this->args['login']) || empty($this->args['apiKey'])) {
+        if (empty($this->args['apiKey'])) {
             $this->error->add(__('Service is not well configured.'));
 
             return false;
         }
 
-        $args = $this->args;
-        $args['hash'] = 'WP9vc';
-        if (!($response = self::post($this->url_api . 'expand', $args, true))) {
+        $args = json_encode(['domain' => 'bit.ly', 'bitlink_id' => 'bit.ly/WP9vc'], JSON_UNESCAPED_SLASHES);
+        if (!($response = self::post($this->url_api . 'expand', $args, true, false, $this->headers()))) {
             $this->error->add(__('Failed to call service.'));
-
-            return false;
-        }
-
-        $rsp = simplexml_load_string($response);
-
-        $err_msg = (string) $rsp->status_txt;
-        if ($err_msg != 'OK') {
-            $err_no = (integer) $rsp->status_code;
-            $this->error->add(sprintf(__('An error occured with code %s and message "%s"'), $err_no, $err_msg));
 
             return false;
         }
@@ -103,29 +74,26 @@ class bitlyKutrlService extends kutrlService
 
     public function createHash($url, $hash = null)
     {
-        $args = array_merge($this->args, ['longUrl' => $url]);
+        $args = json_encode(['domain' => 'bit.ly', 'long_url' => $url]);
 
-        if (!($response = self::post($this->url_api . 'shorten', $args, true))) {
+        if (!($response = self::post($this->url_api . 'shorten', $args, true, false, $this->headers()))) {
             $this->error->add(__('Failed to call service.'));
 
             return false;
         }
 
-        $rsp = simplexml_load_string($response);
-
-        $err_msg = (string) $rsp->status_txt;
-        if ($err_msg != 'OK') {
-            $err_no = (integer) $rsp->status_code;
-            $this->error->add(sprintf(__('An error occured with code %s and message "%s"'), $err_no, $err_msg));
-
-            return false;
-        }
+        $rsp = json_decode($response);
 
         $rs = new ArrayObject();
-        $rs->hash = (string) $rsp->data[0]->hash;
-        $rs->url  = (string) $rsp->data[0]->long_url;
+        $rs->hash = str_replace($this->url_base, '', (string) $rsp->link);
+        $rs->url  = (string) $rsp->long_url;
         $rs->type = $this->id;
 
         return $rs;
+    }
+
+    private function headers()
+    {
+        return ['Authorization: Bearer ' . $this->args['apiKey'], 'Content-Type: application/json'];
     }
 }